Reduced log warnings when connection failed on Bungee (certain requests)
