+++ Plan/src/main/java/com/djrapitops/plan/system/webserver/response/pages/InspectPageResponse.java
@@ -4,6 +4,7 @@ import com.djrapitops.plan.system.settings.theme.Theme;
 import com.djrapitops.plan.system.webserver.response.Response;
 import com.djrapitops.plan.system.webserver.response.cache.PageId;
 import com.djrapitops.plan.system.webserver.response.cache.ResponseCache;
+import com.djrapitops.plan.system.webserver.response.errors.ErrorResponse;
 import com.djrapitops.plan.system.webserver.response.pages.parts.InspectPagePluginsContent;
 import org.apache.commons.lang3.text.StrSubstitutor;
 
@@ -41,4 +42,11 @@ public class InspectPageResponse extends Response {
         return new String[]{"<meta http-equiv=\"refresh\" content=\"5\" /><li><a><i class=\"fa fa-spin fa-refresh\"></i> Calculating...</a></li>", ""};
     }
 
+    public static InspectPageResponse getRefreshing() {
+        ErrorResponse refreshPage = new ErrorResponse();
+        refreshPage.setTitle("Player page request is being processed..");
+        refreshPage.setParagraph("<meta http-equiv=\"refresh\" content=\"2\" /><i class=\"fa fa-refresh fa-spin\" aria-hidden=\"true\"></i> Page will refresh automatically..");
+        refreshPage.replacePlaceholders();
+        return new InspectPageResponse(null, refreshPage.getContent());
+    }
 }
